pyproject.toml
[tool.poetry]
name = "tableai"
version = "0.0.1"
description = "This feature is responsible for chaining together LLM chat responses for RAG PDF table extraction"
authors = ["Schlafdata <john@schlafdata.com>"]
readme = "README.md"
license = ""

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pandas = "*"
pyyaml = "6.0.1"
jinja2 = "3.1.3"
pytest = "8.1.1"
pytest-mock = "3.14.0"
boto3= "1.34.74"
kubernetes = "29.0.0"
psycopg2-binary = "2.9.10"
pydantic = "2.9.2"
black = "24.10.0"
pydantic-settings = "2.6.0"
psutil = "6.1.0"
networkx = "3.4.2"
requests = "*"
json-repair = "0.32.0"
fastapi = {extras = ["all"], version = "^0.115.4"}
PyMuPDF = "1.25.5"
pdf2image = "1.17.0"
numpy = "2.2.4"
pillow = "11.2.1"
scikit-learn = "1.6.1"
scipy = "1.15.2"
dropbox = "12.0.2"
sqlmodel = "0.0.24"
pytesseract = "0.3.13"
ocrmypdf = "16.10.1"
python-dotenv = "1.1.0"
alembic = "1.15.2"
dateparser = "1.2.1"

[tool.poetry.dev-dependencies]
# Add your development dependencies here. For example:
# pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/jschlafdata/llm-migrate-output"